refactor(pages): split out screens & components
