+++ src/screens/home/components/about/styles.scss
@@ -1,0 +1,8 @@
+.container {
+  position: relative;
+  height: 100%;
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
