+++ src/screens/home/components/introduction/index.js
@@ -1,0 +1,14 @@
+import React from "react"
+import './styles.scss'
+
+const Intro = () => {
+  return (
+    <div
+      className='container'
+    >
+      <h1>Vincent Aceto</h1>
+    </div>
+  )
+}
+
+export default Intro
