+++ src/screens/home/components/introduction/styles.scss
@@ -1,0 +1,8 @@
+.container {
+    height: 100%;
+    width: 100%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background-color: rgb(193, 88, 88);
+}
